bugfix area calculation in contsructor was in int i added double casting

+++ src/main/java/org/treathunter/calculation/CalculatorCutoutElement.java
@@ -4,7 +4,7 @@ public class CalculatorCutoutElement {
 	public double area;
 	
 	public CalculatorCutoutElement(int lenght, int width) {
-		area = lenght * width / 1_000_000; //m^2
+		area =(double) lenght * width / 1_000_000; //m^2
 	}
 	
 	public CalculatorCutoutElement(double area) {
